added axolotl animation, made alldays a prefab
