+++ Assets/Materials.meta~HEAD.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4b92188c8206f409688a576331782cfe
+timeCreated: 1492917323
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
