+++ Assets/Materials.meta~master.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: b361476e1bd859346b705863c0d2f754
-timeCreated: 1491951223
+guid: bdaaf1f617fff45fd837c02ae2773d4a
+timeCreated: 1492917323
 licenseType: Free
 DefaultImporter:
   userData:
