+++ Assets/julian1.unity.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0fed3d54dd6e04fa7863e7875252f196
+timeCreated: 1491860720
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
