+++ Assets/prefabs/ALLDAYS.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6d9c3ef76e4f44bb1bef9f69e868c03d
+timeCreated: 1492921676
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
